Valid associations against content schemas
Having to add a title as part of creating the document to test makes it
clear this test is vulnerable to different document types having
different required fields. This is likely to be a global issue, though,
so this code should be refactored when we decide on how to deal with it.