Add a JSON field to documents to store images
The intention is that this will store an array of objects that represent
images rather than using the ActiveStorage::Attachment model.

The reason for doing this is that:

- We need to store more data around images than we have in
ActiveStorage::Attachment
- With our intention that document evolves into a Revision model that
records all changes it seems consistent to store this as JSON rather
than in separate models.
- We still gain almost all of the benefits from ActiveStorage despite
bypassing Attachment as the value seems to lie in the blob model.

I'm not 100% sure this is the right approach but it seems initially to be
most consistent with how we're modelling currently. Might well be that
it transpires this is a bad idea and separate model is better.