Add check for existing paths in the Publishing API
Previously we would only generate base paths that were unique within
Content Publisher, and any conflict with an existing base path in the
Publishing API would result in an exception when creating a preview.

This adds a check to determine if the base path is already reserved in
the Publishing API. Similarly to the check for topics, we silently
ignore any errors from the Publishing API until the user attempts to
manually generate a preview. Because the Publishing API has no direct
endpoint to check path reservations, we use the lookup content id one.

In order to avoid adding the lookup stub to a large number of tests,
this new check is only run if the specified in the format definition.